requirements.txt
python-dotenv
# Жаба Лейнера (aiogram v3)
aiogram>=3.0,<4
APScheduler>=3.10
aiosqlite>=0.19,<0.21
python-dotenv>=1.0
pytz>=2024.1
aiohttp>=3.9
